+++ vite-project/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "axios": "^1.12.2",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.9.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
