+++ vite-project/src/pages/board.jsx
@@ -1,0 +1,5 @@
+const BoardPage = () => {
+  return <div>BoardPage</div>;
+};
+
+export default BoardPage;
